refactors robot URDF to use left/right/back identifies removes Pose2D and adds preliminary GhostRobotState changes encoder angles to floats reverses encoders and changes update rate to 5ms refactors particle_filter_node to ghost_estimator_node refactors ghost_estimator config to have namespaces for particle filter and odometry fixes rear module port change from ESD board removal adds joint state publisher to ghost_estimator confirms working steering joint states

+++ ghost_ros/src/ros_nodes/jetson_v5_serial_node.cpp
@@ -163,7 +163,7 @@ namespace ghost_serial
             encoder_state_msg.encoders[motor_id].device_id = motor_id;
 
             // Copy encoder angle
-            uint32_t angle;
+            float angle;
             memcpy(&angle, buffer + 4 * (buffer_index++), 4);
             encoder_state_msg.encoders[motor_id].current_angle = angle;
 
@@ -180,7 +180,7 @@ namespace ghost_serial
             encoder_state_msg.encoders[sensor_id].device_id = sensor_id;
 
             // Copy encoder angle
-            uint32_t angle;
+            float angle;
             memcpy(&angle, buffer + 4 * (buffer_index++), 4);
             encoder_state_msg.encoders[sensor_id].current_angle = angle;
